pyproject.toml
[project]
name = "crud_task"
version = "v1.0"
description = "Test task for Fablite."
authors = [{name = "Eugene Kovalenko", email = "eugene.kovalenko.work@gmail.com"}]
readme = "README.md"
requires-python = ">= 3.12"
dependencies = [
  "djangorestframework",
  "djangorestframework-simplejwt",
  "python-dotenv",
]
